ignore proxy-default.el file
This file contains the current proxy settings, which can vary from site to site.
But a given site can configure, according to the network environment, different
settings in different moments. For example, the Indra's laptop can be working
inside Indra's premises with the official proxy, but also outside without a
proxy.